count reviews properly
A few things here. First, the github api returns reviews in
chronological order, so we have to reverse them to get the
most recent reviews.

Also, we need to exclude older reviews from that chronological
list for reviewers that have a more recent review.

Finally, we just need to count the approvals.